Add expected tests

+++ DSE/samples/primitives/overflow.go
@@ -2,12 +2,12 @@ package main
 
 type Overflow struct{}
 
-func (o *Overflow) ShortOverflow(x int16, y int) int {
+func (o *Overflow) ShortOverflow(x int16, y int16) int16 {
 	if y > 10 || y <= 0 {
 		return 0
 	}
-	if int(x)+y < 0 && x > 0 {
+	if x+y < 0 && x > 0 {
 		panic("IllegalStateException")
 	}
-	return int(x) + y
+	return x + y
 }
